Add modular feature control and dynamic feature status identifier

+++ routes/web/auth.php
@@ -1,5 +1,7 @@
 <?php
 
+use App\Features\AppFeatures;
+use App\Features\AuthFeatures;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Auth\ApiTokenController;
 use App\Http\Controllers\Auth\PasswordController;
@@ -25,20 +27,28 @@ Route::group([
     'middleware' => 'guest',
 ], function (): void {
     // Sign In Routes...
-    Route::get('/signin', [AuthenticatedSessionController::class, 'create'])->name('signin');
-    Route::post('/signin', [AuthenticatedSessionController::class, 'store']);
-    Route::get('/tfa-challenge', [TwoFactorAuthenticatedSessionController::class, 'create'])->name('tfa.signin');
-    Route::post('/tfa-challenge', [TwoFactorAuthenticatedSessionController::class, 'store']);
+    if (AuthFeatures::signin()) {
+        Route::get('/signin', [AuthenticatedSessionController::class, 'create'])->name('signin');
+        Route::post('/signin', [AuthenticatedSessionController::class, 'store']);
+        if (AuthFeatures::hasTwoFactorAuthentication()) {
+            Route::get('/tfa-challenge', [TwoFactorAuthenticatedSessionController::class, 'create'])->name('tfa.signin');
+            Route::post('/tfa-challenge', [TwoFactorAuthenticatedSessionController::class, 'store']);
+        }
+    }
 
     // Sign Up Routes...
-    Route::get('/signup', [RegisteredUserController::class, 'create'])->name('signup');
-    Route::post('/signup', [RegisteredUserController::class, 'store']);
+    if (AuthFeatures::signin() && AuthFeatures::signup()) {
+        Route::get('/signup', [RegisteredUserController::class, 'create'])->name('signup');
+        Route::post('/signup', [RegisteredUserController::class, 'store']);
+    }
 
     // Reset Password Routes...
-    Route::get('/forgot-password', [PasswordResetLinkController::class, 'create'])->name('password.request');
-    Route::post('/forgot-password', [PasswordResetLinkController::class, 'store'])->name('password.email');
-    Route::get('/reset-password/{token}', [NewPasswordController::class, 'create'])->name('password.reset');
-    Route::post('/reset-password', [NewPasswordController::class, 'store'])->name('password.update');
+    if (AuthFeatures::passwordReset()) {
+        Route::get('/forgot-password', [PasswordResetLinkController::class, 'create'])->name('password.request');
+        Route::post('/forgot-password', [PasswordResetLinkController::class, 'store'])->name('password.email');
+        Route::get('/reset-password/{token}', [NewPasswordController::class, 'create'])->name('password.reset');
+        Route::post('/reset-password', [NewPasswordController::class, 'store'])->name('password.update');
+    }
 });
 
 Route::group([
@@ -48,48 +58,62 @@ Route::group([
     Route::post('/signout', [AuthenticatedSessionController::class, 'destroy'])->name('signout');
 
     // Email Verification Routes...
-    Route::get('/email/verify', [EmailVerificationPromptController::class, '__invoke'])->name('verification.notice');
-    Route::get('/email/verify/{id}/{hash}', [VerifyEmailController::class, '__invoke'])
-        ->middleware(['signed', 'throttle:6,1'])
-        ->name('verification.verify');
-    Route::post('/email/verification-notification', [EmailVerificationNotificationController::class, 'store'])
-        ->middleware(['throttle:6,1'])
-        ->name('verification.send');
+    if (AuthFeatures::emailVerification()) {
+        Route::get('/email/verify', [EmailVerificationPromptController::class, '__invoke'])->name('verification.notice');
+        Route::get('/email/verify/{id}/{hash}', [VerifyEmailController::class, '__invoke'])
+            ->middleware(['signed', 'throttle:6,1'])
+            ->name('verification.verify');
+        Route::post('/email/verification-notification', [EmailVerificationNotificationController::class, 'store'])
+            ->middleware(['throttle:6,1'])
+            ->name('verification.send');
+    }
 
     Route::group([
         'prefix' => 'user',
     ], function (): void {
         // Update & Confirm Password Routes...
         Route::put('/password', [PasswordController::class, 'update'])->name('user-password.update');
-        Route::get('/confirm-password', [ConfirmablePasswordController::class, 'show'])->name('password.confirm');
-        Route::post('/confirm-password', [ConfirmablePasswordController::class, 'store']);
-        Route::get('/confirmed-password-status', [ConfirmedPasswordStatusController::class, 'show'])->name('password.confirmation');
+        if (AuthFeatures::passwordConfirmation()) {
+            Route::get('/confirm-password', [ConfirmablePasswordController::class, 'show'])->name('password.confirm');
+            Route::post('/confirm-password', [ConfirmablePasswordController::class, 'store']);
+            Route::get('/confirmed-password-status', [ConfirmedPasswordStatusController::class, 'show'])->name('password.confirmation');
+        }
 
         // Other Browser Sessions Route...
-        Route::delete('/other-browser-sessions', [OtherBrowserSessionsController::class, 'destroy'])->name('other-browser-sessions.destroy');
+        if (AuthFeatures::otherBrowserSessions()) {
+            Route::delete('/other-browser-sessions', [OtherBrowserSessionsController::class, 'destroy'])->name('other-browser-sessions.destroy');
+        }
 
         // User Profile Routes...
         Route::get('/profile', [UserProfileController::class, 'show'])->name('profile.show');
         Route::put('/profile', [UserProfileController::class, 'update'])->name('profile.update');
-        Route::delete('/profile-photo', [UserProfilePhotoController::class, 'destroy'])->name('user-photo.destroy');
         Route::delete('/', [CurrentUserController::class, 'destroy'])->name('user.destroy');
 
+        // User Profile Picture...
+        if (AppFeatures::hasProfilePhoto()) {
+            Route::delete('/profile-photo', [UserProfilePhotoController::class, 'destroy'])->name('user-photo.destroy');
+        }
+
         // API Token Routes...
-        Route::get('/api-tokens', [ApiTokenController::class, 'index'])->name('api-tokens.index');
-        Route::post('/api-tokens', [ApiTokenController::class, 'store'])->name('api-tokens.store');
-        Route::put('/api-tokens/{token}', [ApiTokenController::class, 'update'])->name('api-tokens.update');
-        Route::delete('/api-tokens/{token}', [ApiTokenController::class, 'destroy'])->name('api-tokens.destroy');
+        if (AppFeatures::hasApiToken()) {
+            Route::get('/api-tokens', [ApiTokenController::class, 'index'])->name('api-tokens.index');
+            Route::post('/api-tokens', [ApiTokenController::class, 'store'])->name('api-tokens.store');
+            Route::put('/api-tokens/{token}', [ApiTokenController::class, 'update'])->name('api-tokens.update');
+            Route::delete('/api-tokens/{token}', [ApiTokenController::class, 'destroy'])->name('api-tokens.destroy');
+        }
     });
 
-    Route::group([
-        'prefix' => 'user',
-        'middleware' => config('auth.password_confirmation') ? ['auth', 'password.confirm'] : ['auth'],
-    ], function (): void {
-        // Two Factor Authentication Routes...
-        Route::post('/tfa', [TwoFactorAuthenticationController::class, 'store']);
-        Route::delete('/tfa', [TwoFactorAuthenticationController::class, 'destroy']);
-        Route::get('/tfa-qr-code', [TwoFactorQrCodeController::class, 'show']);
-        Route::get('/tfa-recovery-codes', [RecoveryCodeController::class, 'index']);
-        Route::post('/tfa-recovery-codes', [RecoveryCodeController::class, 'store']);
-    });
+    if (AuthFeatures::hasTwoFactorAuthentication()) {
+        Route::group([
+            'prefix' => 'user',
+            'middleware' => AuthFeatures::hasPasswordConfirmation() ? ['auth', 'password.confirm'] : ['auth'],
+        ], function (): void {
+            // Two Factor Authentication Routes...
+            Route::post('/tfa', [TwoFactorAuthenticationController::class, 'store']);
+            Route::delete('/tfa', [TwoFactorAuthenticationController::class, 'destroy']);
+            Route::get('/tfa-qr-code', [TwoFactorQrCodeController::class, 'show']);
+            Route::get('/tfa-recovery-codes', [RecoveryCodeController::class, 'index']);
+            Route::post('/tfa-recovery-codes', [RecoveryCodeController::class, 'store']);
+        });
+    }
 });
